\name{cancerdata-package}
\alias{cancerdata}
\docType{package}

\title{
Development and validation of diagnostic tests from high-dimensional molecular data: Datasets
}

\description{
This package contains dataset for the R package cancerclass.
}

\details{
\tabular{ll}{
Package: \tab cancerdata\cr
Type: \tab Package\cr
Version: \tab 1.1.0 \cr
Date: \tab 2010-10-26\cr
License: \tab GPL (>=2)
}}

\author{
Jan Budczies \email{jan.budczies@charite.de},
Daniel Kosztyla \email{danielkossi@hotmail.com}
}

\references{
[1] Michiels S, Koscielny S, Hill C (2005), \emph{Prediction of cancer outcome
with microarrays: a multiple random validation strategy}, Lancet 365:488-492.
}

\seealso{ 
  \code{\link{VEER1}}
}  

\examples{
### see: help(VEER1);
}

\keyword{ package }
